+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - 🛠️ Adjust to compatible camera dependency to prevent face not found
 - 🔐 Ajdust threshold for smile and look down challenge
+- 🎨 Add showCurrentStep parameter (default : false)
 
 ## 1.0.2 🚀